requirements.txt
matplotlib==3.8.0
numpy==1.26.1
openpyxl==3.1.2
pandas==2.1.2
pygad==3.2.0
PyYAML==6.0.1